pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "homing"
version = "0.1.0"
description = "Make a personal laptop legible to agents by walking $HOME and producing a structured ~/system/ representation."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
authors = [
    { name = "homing contributors" },
]
keywords = ["agent", "filesystem", "classification", "migration", "home"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Filesystems",
    "Topic :: Utilities",
]
dependencies = [
    "typer[all]>=0.12",
    "pydantic>=2.0",
    "pyyaml>=6",
    "python-frontmatter>=1.0",
    "anthropic>=0.30",
    "rich>=13",
    # cabinet (sibling package): personal-document triage
    "Pillow>=10",
    "pdf2image>=1.17",
    "pypdf>=4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-cov>=5",
    "ruff>=0.5",
    "black>=24",
    "mypy>=1.10",
]

[project.scripts]
homing = "homing.cli:app"
cabinet = "cabinet.cli:app"

[project.urls]
Homepage = "https://github.com/itisaevalex/homing"
Issues = "https://github.com/itisaevalex/homing/issues"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
homing = ["py.typed"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra"
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["homing"]

[tool.coverage.report]
show_missing = true
skip_covered = false

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "TID"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
no_implicit_optional = true
